rename sidecars package to linkedimage
